move query_result.rs to new response mod
