+++ scylla/src/lib.rs
@@ -132,10 +132,10 @@ pub use frame::response::cql_to_rust;
 pub use frame::response::cql_to_rust::FromRow;
 
 pub use execution::ExecutionProfile;
+pub use response::{IntoTypedRows, QueryResult};
 pub use session::CachingSession;
 pub use session::SessionBuilder;
 pub use session::{Session, SessionConfig};
-pub use transport::query_result::{IntoTypedRows, QueryResult};
 
 #[cfg(feature = "cloud")]
 pub use session::CloudSessionBuilder;
