+++ scylla/src/response/mod.rs
@@ -1,2 +1,8 @@
 mod query_response;
+mod query_result;
+
 pub(crate) use query_response::{NonErrorQueryResponse, QueryResponse};
+pub use query_result::{
+    FirstRowError, FirstRowTypedError, IntoTypedRows, MaybeFirstRowTypedError, QueryResult,
+    RowsExpectedError, RowsNotExpectedError, SingleRowError, SingleRowTypedError, TypedRowIter,
+};
