+++ scylla/src/response/query_result.rs
