+++ scylla/src/tests/cql_types_test.rs
@@ -5,9 +5,9 @@ use crate::frame::value::Counter;
 use crate::frame::value::Value;
 use crate::frame::value::{Date, Time, Timestamp};
 use crate::macros::{FromUserType, IntoUserType};
+use crate::response::IntoTypedRows;
 use crate::session::Session;
 use crate::test_utils::create_new_session_builder;
-use crate::transport::query_result::IntoTypedRows;
 use crate::utils::test_utils::unique_keyspace_name;
 use bigdecimal::BigDecimal;
 use chrono::{Duration, NaiveDate};
