+++ scylla/src/transport/mod.rs
@@ -1,3 +1,1 @@
-pub mod query_result;
-
 pub use crate::frame::{Authenticator, Compression};
